[refactor] extract and reuse buildDiagnosisKey helper methods
