+++ common/persistence/src/test/java/app/coronawarn/server/common/persistence/domain/DiagnosisKeyBuilderTest.java
@@ -21,9 +21,11 @@
 package app.coronawarn.server.common.persistence.domain;
 
 import static app.coronawarn.server.common.persistence.domain.validation.ValidSubmissionTimestampValidator.SECONDS_PER_HOUR;
+import static app.coronawarn.server.common.persistence.service.DiagnosisKeyServiceTestHelper.buildDiagnosisKeyForSubmissionTimestamp;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.catchThrowable;
+
 import app.coronawarn.server.common.persistence.exception.InvalidDiagnosisKeyException;
 import app.coronawarn.server.common.protocols.external.exposurenotification.TemporaryExposureKey;
 import com.google.protobuf.ByteString;
@@ -32,7 +34,6 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneOffset;
-import java.time.temporal.ChronoUnit;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -195,26 +196,26 @@ class DiagnosisKeyBuilderTest {
   @ValueSource(longs = {-1L, Long.MAX_VALUE})
   void submissionTimestampMustBeValid(long submissionTimestamp) {
     assertThat(
-        catchThrowable(() -> keyWithSubmissionTimestamp(submissionTimestamp)))
+        catchThrowable(() -> buildDiagnosisKeyForSubmissionTimestamp(submissionTimestamp)))
         .isInstanceOf(InvalidDiagnosisKeyException.class);
   }
 
   @Test
   void submissionTimestampMustNotBeInTheFuture() {
     assertThat(catchThrowable(
-        () -> keyWithSubmissionTimestamp(getCurrentHoursSinceEpoch() + 1)))
+        () -> buildDiagnosisKeyForSubmissionTimestamp(getCurrentHoursSinceEpoch() + 1)))
             .isInstanceOf(InvalidDiagnosisKeyException.class);
-    assertThat(catchThrowable(() -> keyWithSubmissionTimestamp(
+    assertThat(catchThrowable(() -> buildDiagnosisKeyForSubmissionTimestamp(
         Instant.now().getEpochSecond() /* accidentally forgot to divide by SECONDS_PER_HOUR */)))
             .isInstanceOf(InvalidDiagnosisKeyException.class);
   }
 
   @Test
   void submissionTimestampDoesNotThrowOnValid() {
-    assertThatCode(() -> keyWithSubmissionTimestamp(0L)).doesNotThrowAnyException();
-    assertThatCode(() -> keyWithSubmissionTimestamp(getCurrentHoursSinceEpoch())).doesNotThrowAnyException();
+    assertThatCode(() -> buildDiagnosisKeyForSubmissionTimestamp(0L)).doesNotThrowAnyException();
+    assertThatCode(() -> buildDiagnosisKeyForSubmissionTimestamp(getCurrentHoursSinceEpoch())).doesNotThrowAnyException();
     assertThatCode(
-        () -> keyWithSubmissionTimestamp(Instant.now().minus(Duration.ofHours(2)).getEpochSecond() / SECONDS_PER_HOUR))
+        () -> buildDiagnosisKeyForSubmissionTimestamp(Instant.now().minus(Duration.ofHours(2)).getEpochSecond() / SECONDS_PER_HOUR))
             .doesNotThrowAnyException();
   }
   
@@ -245,14 +246,6 @@ class DiagnosisKeyBuilderTest {
         .withKeyData(expKeyData)
         .withRollingStartIntervalNumber(expRollingStartIntervalNumber)
         .withTransmissionRiskLevel(expTransmissionRiskLevel).build();
-  }
-
-  private DiagnosisKey keyWithSubmissionTimestamp(long submissionTimestamp) {
-    return DiagnosisKey.builder()
-        .withKeyData(expKeyData)
-        .withRollingStartIntervalNumber(expRollingStartIntervalNumber)
-        .withTransmissionRiskLevel(expTransmissionRiskLevel)
-        .withSubmissionTimestamp(submissionTimestamp).build();
   }
 
   private void assertDiagnosisKeyEquals(DiagnosisKey actDiagnosisKey) {
